Moved get_verbose_name to utils.
