+++ request/plugins.py
@@ -1,19 +1,11 @@
 # -*- coding: utf-8 -*-
-import re
-
 from django.db.models import Count
 from django.template.loader import render_to_string
 
 from . import settings
 from .models import Request
 from .traffic import modules
-
-
-def get_verbose_name(class_name):
-    '''
-    Calculate the verbose_name by converting from InitialCaps to "lowercase with spaces".
-    '''
-    return re.sub('(((?<=[a-z])[A-Z])|([A-Z](?![A-Z]|$)))', ' \\1', class_name).strip()
+from .utils import get_verbose_name
 
 
 def set_count(items):
